+++ src/processx.h
@@ -46,7 +46,7 @@ SEXP processx_exec(SEXP command, SEXP args, SEXP pty, SEXP pty_options,
 		   SEXP connections, SEXP env, SEXP windows_verbatim_args,
 		   SEXP windows_hide_window, SEXP windows_detached_process,
 		   SEXP private_, SEXP cleanup, SEXP wd, SEXP encoding,
-		   SEXP tree_id);
+		   SEXP tree_id, SEXP linux_pdeathsig);
 SEXP processx_wait(SEXP status, SEXP timeout, SEXP name);
 SEXP processx_is_alive(SEXP status, SEXP name);
 SEXP processx_pty_close(SEXP status, SEXP name);
@@ -115,6 +115,7 @@ typedef struct {
   int pty_echo;
   int pty_rows;
   int pty_cols;
+  int linux_pdeathsig;
 } processx_options_t;
 
 #ifdef __cplusplus